requirements.txt
altgraph==0.17
click==7.1.2
cycler==0.10.0
future==0.18.2
kiwisolver==1.2.0
matplotlib==3.2.1
minimalmodbus==1.0.2
numpy==1.18.4
pandas==0.24.2
pefile==2019.4.18
PyInstaller==3.6
pyparsing==2.4.7
PyQt5==5.13.0
PyQt5-sip==12.7.2
pyserial==3.4
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
pywin32-ctypes==0.2.0
six==1.15.0
